Skips extensions which already exists

+++ index.js
@@ -3,7 +3,10 @@ const path = require('path');
 
 const Promise = require('bluebird');
 
+
 const transformers = {
+  // All transformers take an Uint8Array as input and output
+  // an object. The result object will include the transformed buffer
   eot: require('ttf2eot'),
   woff: require('ttf2woff'),
   woff2: require('ttf2woff2'),
@@ -19,8 +22,15 @@ const writeTransformedFonts = (dirname, filename, ttfBuffer) => {
   const writeFile = Promise.promisify(fs.writeFile);
 
   const promises = Object.keys(transformers).map((ext) => {
-    const buffer = transformFontTo(ext, ttfBuffer);
     const filepath = `./${dirname}/${filename}.${ext}`;
+
+    // Skips transform and write actions when a font file already exists
+    const exists = fs.existsSync(filepath);
+    if (exists === true) {
+      return Promise.resolve();
+    }
+
+    const buffer = transformFontTo(ext, ttfBuffer);
     return writeFile(filepath, buffer);
   });
 
@@ -53,3 +63,7 @@ const fontfeest = function fontfeest(file, fontFamily, fontWeight = 400, fontSty
 };
 
 module.exports = fontfeest;
+
+fontfeest('./fixtures/OpenSans-Light.ttf', 'Open Sans', 300).then((css) => {
+  console.log(css);
+});
